+++ R/tbl_hierarchical_rate_by_grade.R
@@ -453,8 +453,11 @@ tbl_hierarchical_rate_by_grade <- function(data,
   grade <- dplyr::last(variables)
 
   if (!is_empty(grade_groups)) {
+    keep_grade <- vapply(grade_groups, function(x) {
+      any(x %in% levels(data[[grade]]))
+    }, TRUE)
     # if any grade groups present, replace grades with their grade groups
-    data[[grade]] <- do.call(fct_collapse, args = c(list(f = data[[grade]]), grade_groups))
+    data[[grade]] <- do.call(fct_collapse, args = c(list(f = data[[grade]]), grade_groups[keep_grade]))
   }
 
   # move grade variable to `by` to get rates by highest grade